requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
requests>=2.26.0
solders>=0.9.0
click>=8.1.0
rich>=10.0.0 